+++ app/lib/mappers/mapService.js
@@ -18,7 +18,9 @@ function mapService(rawService) {
     id: utils.getId(entry.id),
     location: mapLocation(service),
     name: service.name._,
-    openingTimes: mapNonCore.openingTimes(service.nonCoreElements),
+    openingTimes: {
+      description: mapNonCore.openingTimes(service.nonCoreElements),
+    },
     serviceDetails: mapNonCore.serviceDetails(service.nonCoreElements),
     serviceType: service.type._,
     venueType: mapNonCore.venueType(service.nonCoreElements),