+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add GSD ID to output data
 - Map telephone number to `contacts.telephoneNumber`
 - Map service type to `serviceType`
+- Map opening times to `openingTimes.description`
 
 0.2.0 / 2018-03-28
 ==================